Link up localisation to API requests
