+++ app/src/main/java/com/hrb116/waterlogged/tile/addwater/AddWaterLayout.kt
@@ -21,6 +21,7 @@ import com.hrb116.waterlogged.R
 import com.hrb116.waterlogged.tile.previewResources
 import com.google.android.horologist.compose.tools.LayoutRootPreview
 import com.google.android.horologist.compose.tools.buildDeviceParameters
+import com.hrb116.waterlogged.tools.getLocalisedWaterVolume
 import com.hrb116.waterlogged.tools.getWaterUnit
 
 fun addWaterLayout(context: Context, container: String): LayoutElementBuilders.LayoutElement {
@@ -71,30 +72,6 @@ fun addWaterLayout(context: Context, container: String): LayoutElementBuilders.L
         .setContent(contentChip)
         .setPrimaryChipContent(primaryChip)
         .build()
-}
-
-private fun getLocalisedWaterVolume(context: Context, container: String): String {
-    val unit = getWaterUnit(context) ?: "ml"
-    if (unit == "ml") {
-        return when (container) {
-            "glass" -> "250ml"
-            "bottle" -> "500ml"
-            else -> "750ml"
-        }
-    } else if (unit == "fl oz") {
-        return when (container) {
-            "glass" -> "8oz"
-            "bottle" -> "16oz"
-            else -> "24oz"
-        }
-    } else {
-        // cup
-        return when (container) {
-            "glass" -> "1 cup"
-            "bottle" -> "2 cups"
-            else -> "3 cups"
-        }
-    }
 }
 
 @Preview(device = WearDevices.SMALL_ROUND)
